@@ -7,7 +7,7 @@ name = "locust"
 license = { text = "MIT" }
 description = "Developer friendly load testing framework"
 dynamic = ["version"]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "gevent >=22.10.2",
     "flask >=2.0.0",
@@ -15,7 +15,7 @@ dependencies = [
     "requests >=2.26.0",
     "msgpack >=1.0.0",
     "pyzmq >=25.0.0",
-    "geventhttpclient >=2.0.11",
+    "geventhttpclient ==2.2.1",
     "ConfigArgParse >=1.5.5",
     "tomli >=1.1.0; python_version<'3.11'",
     "psutil >=5.9.1",
@@ -31,7 +31,6 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -61,7 +60,7 @@ license-files = ["LICENSE"]
 include-package-data = false
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 120
 extend-exclude = [
     "build",
@@ -73,7 +72,6 @@ lint.select = ["E", "F", "W", "UP", "FA102", "I001"]
 
 [tool.ruff.lint.per-file-ignores]
 "examples/*" = ["F841"]
-"locust/test/*" = ["F841"]
 
 [tool.ruff.lint.isort]
 section-order = ["future", "locust", "standard-library", "third-party", "first-party", "local-folder"]
@@ -104,8 +102,15 @@ locust = "locust.main:main"
 [tool.mypy]
 # missing type stubs
 ignore_missing_imports = true
-python_version = "3.8"
+python_version = "3.9"
 
 [[tool.mypy.overrides]]
 module = ["locust.dispatch"]
 no_strict_optional = true
+
+[tool.pyright]
+exclude = [
+    "locust/test",
+]
+reportOptionalMemberAccess = "none"
+analyzeUnannotatedFunctions = false
