@@ -847,6 +847,75 @@ class TestRampUpUsersFromZero(unittest.TestCase):
         delta = time.perf_counter() - ts
         self.assertTrue(0 <= delta <= _TOLERANCE, delta)
 
+    def test_implementation_of_dispatch_distribution_with_gcd(self):
+        class User1(User):
+            weight = 4
+
+        class User2(User):
+            weight = 5
+
+        user_classes = [User1, User2]
+        worker_node1 = WorkerNode("1")
+
+        sleep_time = 0.2  # Speed-up test
+
+        users_dispatcher = UsersDispatcher(worker_nodes=[worker_node1], user_classes=user_classes)
+        users_dispatcher.new_dispatch(target_user_count=9, spawn_rate=9)
+
+        users_dispatcher._wait_between_dispatch = sleep_time
+
+        ts = time.perf_counter()
+        self.assertDictEqual(
+            next(users_dispatcher),
+            {
+                "1": {"User1": 4, "User2": 5},
+            },
+        )
+        delta = time.perf_counter() - ts
+        self.assertTrue(0 <= delta <= _TOLERANCE, delta)
+
+        ts = time.perf_counter()
+        self.assertRaises(StopIteration, lambda: next(users_dispatcher))
+        delta = time.perf_counter() - ts
+        self.assertTrue(0 <= delta <= _TOLERANCE, delta)
+
+    def test_implementation_of_dispatch_distribution_with_gcd_float_weight(self):
+        class User1(User):
+            weight = 0.8
+
+        class User2(User):
+            weight = 1
+
+        normalized_weights_to_min_int = 5  # User1: 0.8 * 5 = 4; User2: 1 * 5 = 5
+
+        user_classes = [User1, User2]
+        worker_node1 = WorkerNode("1")
+
+        sleep_time = 0.2  # Speed-up test
+
+        users_dispatcher = UsersDispatcher(worker_nodes=[worker_node1], user_classes=user_classes)
+        users_dispatcher.new_dispatch(target_user_count=18, spawn_rate=18)
+
+        users_dispatcher._wait_between_dispatch = sleep_time
+
+        ts = time.perf_counter()
+        self.assertDictEqual(
+            next(users_dispatcher),
+            {
+                "1": {
+                    "User1": int(normalized_weights_to_min_int * User1.weight * 2),
+                    "User2": int(normalized_weights_to_min_int * User2.weight * 2),
+                },
+            },
+        )
+        delta = time.perf_counter() - ts
+        self.assertTrue(0 <= delta <= _TOLERANCE, delta)
+
+        ts = time.perf_counter()
+        self.assertRaises(StopIteration, lambda: next(users_dispatcher))
+        delta = time.perf_counter() - ts
+        self.assertTrue(0 <= delta <= _TOLERANCE, delta)
+
 
 class TestWaitBetweenDispatch(unittest.TestCase):
     def test_wait_between_dispatch(self):
@@ -2098,9 +2167,7 @@ class TestLargeScale(unittest.TestCase):
                     dispatch_iteration_duration <= tol
                     for dispatch_iteration_duration in users_dispatcher.dispatch_iteration_durations
                 ),
-                "One or more dispatch took more than {:.0f}s to compute (max = {}ms)".format(
-                    tol * 1000, 1000 * max(users_dispatcher.dispatch_iteration_durations)
-                ),
+                f"One or more dispatch took more than {tol * 1000:.0f}s to compute (max = {1000 * max(users_dispatcher.dispatch_iteration_durations)}ms)",
             )
 
             self.assertEqual(_user_count(all_dispatched_users[-1]), target_user_count)
@@ -2112,9 +2179,7 @@ class TestLargeScale(unittest.TestCase):
                 self.assertLessEqual(
                     max(user_count_on_workers) - min(user_count_on_workers),
                     1,
-                    "One or more workers have too much users compared to the other workers when user count is {}".format(
-                        _user_count(dispatch_users)
-                    ),
+                    f"One or more workers have too much users compared to the other workers when user count is {_user_count(dispatch_users)}",
                 )
 
             for i, dispatch_users in enumerate(all_dispatched_users):
@@ -2133,9 +2198,7 @@ class TestLargeScale(unittest.TestCase):
                     self.assertLessEqual(
                         error_percent,
                         tol,
-                        "Distribution for user class {} is off by more than {}% when user count is {}".format(
-                            user_class, tol, _user_count(dispatch_users)
-                        ),
+                        f"Distribution for user class {user_class} is off by more than {tol}% when user count is {_user_count(dispatch_users)}",
                     )
 
     def test_ramp_down_from_100_000_to_0_users_with_50_user_classes_and_1000_workers_and_5000_spawn_rate(self):
@@ -2167,9 +2230,7 @@ class TestLargeScale(unittest.TestCase):
                     dispatch_iteration_duration <= tol
                     for dispatch_iteration_duration in users_dispatcher.dispatch_iteration_durations
                 ),
-                "One or more dispatch took more than {:.0f}ms to compute (max = {}ms)".format(
-                    tol * 1000, 1000 * max(users_dispatcher.dispatch_iteration_durations)
-                ),
+                f"One or more dispatch took more than {tol * 1000:.0f}ms to compute (max = {1000 * max(users_dispatcher.dispatch_iteration_durations)}ms)",
             )
 
             self.assertEqual(_user_count(all_dispatched_users[-1]), 0)
@@ -2181,9 +2242,7 @@ class TestLargeScale(unittest.TestCase):
                 self.assertLessEqual(
                     max(user_count_on_workers) - min(user_count_on_workers),
                     1,
-                    "One or more workers have too much users compared to the other workers when user count is {}".format(
-                        _user_count(dispatch_users)
-                    ),
+                    f"One or more workers have too much users compared to the other workers when user count is {_user_count(dispatch_users)}",
                 )
 
             for dispatch_users in all_dispatched_users[:-1]:
@@ -2198,9 +2257,7 @@ class TestLargeScale(unittest.TestCase):
                     self.assertLessEqual(
                         error_percent,
                         tol,
-                        "Distribution for user class {} is off by more than {}% when user count is {}".format(
-                            user_class, tol, _user_count(dispatch_users)
-                        ),
+                        f"Distribution for user class {user_class} is off by more than {tol}% when user count is {_user_count(dispatch_users)}",
                     )
 
 
@@ -3379,9 +3436,7 @@ class TestRampUpUsersFromZeroWithFixed(unittest.TestCase):
             self.target_user_count = target_user_count
 
         def __str__(self):
-            return "<RampUpCase fixed_counts={} weights={} target_user_count={}>".format(
-                self.fixed_counts, self.weights, self.target_user_count
-            )
+            return f"<RampUpCase fixed_counts={self.fixed_counts} weights={self.weights} target_user_count={self.target_user_count}>"
 
     def case_handler(self, cases: list[RampUpCase], expected: list[dict[str, int]], user_classes: list[type[User]]):
         self.assertEqual(len(cases), len(expected))
@@ -3665,8 +3720,6 @@ class TestRampUpDifferentUsers(unittest.TestCase):
 
         worker_node1 = WorkerNode("1")
 
-        sleep_time = 0.2
-
         user_dispatcher = UsersDispatcher(worker_nodes=[worker_node1], user_classes=[User1, User2, User3])
 
         user_dispatcher.new_dispatch(target_user_count=3, spawn_rate=3)
@@ -3692,8 +3745,6 @@ class TestRampUpDifferentUsers(unittest.TestCase):
 
         worker_node1 = WorkerNode("1")
 
-        sleep_time = 0.2
-
         user_dispatcher = UsersDispatcher(worker_nodes=[worker_node1], user_classes=[User1, User2, User3])
 
         user_dispatcher.new_dispatch(target_user_count=10, spawn_rate=10, user_classes=[User2])
@@ -3710,8 +3761,6 @@ class TestRampUpDifferentUsers(unittest.TestCase):
             weight = 1
 
         worker_node1 = WorkerNode("1")
-
-        sleep_time = 0.2
 
         user_dispatcher = UsersDispatcher(worker_nodes=[worker_node1], user_classes=[User1, User2, User3])
 
@@ -3732,8 +3781,6 @@ class TestRampUpDifferentUsers(unittest.TestCase):
             weight = 1
 
         worker_node1 = WorkerNode("1")
-
-        sleep_time = 0.2
 
         user_dispatcher = UsersDispatcher(worker_nodes=[worker_node1], user_classes=[User1, User2, User3])
 
@@ -3756,8 +3803,6 @@ class TestRampUpDifferentUsers(unittest.TestCase):
         worker_node1 = WorkerNode("1")
         worker_node2 = WorkerNode("2")
         worker_node3 = WorkerNode("3")
-
-        sleep_time = 0.2
 
         user_dispatcher = UsersDispatcher(
             worker_nodes=[worker_node1, worker_node2, worker_node3], user_classes=[User1, User2, User3]
@@ -3816,8 +3861,6 @@ class TestRampUpDifferentUsers(unittest.TestCase):
         worker_node1 = WorkerNode("1")
         worker_node2 = WorkerNode("2")
         worker_node3 = WorkerNode("3")
-
-        sleep_time = 0.2
 
         user_dispatcher = UsersDispatcher(
             worker_nodes=[worker_node1, worker_node2, worker_node3], user_classes=[User1, User2, User3]
@@ -4086,7 +4129,7 @@ def _aggregate_dispatched_users(d: dict[str, dict[str, int]]) -> dict[str, int]:
 
 
 def _user_count(d: dict[str, dict[str, int]]) -> int:
-    return sum(map(sum, map(dict.values, d.values())))  # type: ignore
+    return sum(map(sum, map(dict.values, d.values())))
 
 
 def _user_count_on_worker(d: dict[str, dict[str, int]], worker_node_id: str) -> int:
