@@ -3,8 +3,9 @@ from __future__ import annotations
 import logging
 import time
 import traceback
+from collections.abc import Generator
 from contextlib import contextmanager
-from typing import Any, Generator
+from typing import Any
 
 from . import log
 from .exception import InterruptTaskSet, RescheduleTask, RescheduleTaskImmediately, StopUser
