@@ -41,8 +41,7 @@ def get_html_report(
     start_ts = stats.start_time
     start_time = datetime.datetime.utcfromtimestamp(start_ts).strftime("%Y-%m-%d %H:%M:%S")
 
-    end_ts = stats.last_request_timestamp
-    if end_ts:
+    if end_ts := stats.last_request_timestamp:
         end_time = datetime.datetime.utcfromtimestamp(end_ts).strftime("%Y-%m-%d %H:%M:%S")
     else:
         end_time = start_time
