@@ -748,19 +748,19 @@ class TestCatchResponse(WebserverTestCase):
         self.assertEqual(1, self.num_failures)
         self.assertEqual(0, self.num_success)
 
-        with self.locust.client.get("/ultra_fast", catch_response=True) as response:
+        with self.locust.client.get("/ultra_fast", catch_response=True):
             pass
         self.assertEqual(1, self.num_failures)
         self.assertEqual(1, self.num_success)
 
-        with self.locust.client.get("/ultra_fast", catch_response=True) as response:
+        with self.locust.client.get("/ultra_fast", catch_response=True):
             raise ResponseError("Not working")
 
         self.assertEqual(2, self.num_failures)
         self.assertEqual(1, self.num_success)
 
     def test_catch_response_http_fail(self):
-        with self.locust.client.get("/fail", catch_response=True) as response:
+        with self.locust.client.get("/fail", catch_response=True):
             pass
         self.assertEqual(1, self.num_failures)
         self.assertEqual(0, self.num_success)
@@ -793,7 +793,7 @@ class TestCatchResponse(WebserverTestCase):
         class MyTaskSet(TaskSet):
             @task
             def interrupted_task(self):
-                with self.client.get("/ultra_fast", catch_response=True) as r:
+                with self.client.get("/ultra_fast", catch_response=True):
                     raise InterruptTaskSet()
 
         class MyUser(HttpUser):
