@@ -68,6 +68,8 @@ class TestParser(unittest.TestCase):
             web-port = 45787
             headless = true
             tags = ["Critical", "Normal"]
+            [tool.something_else]
+            this = "should be ignored by locust"
             """
 
             file.write(config_data)
@@ -248,7 +250,7 @@ class TestArgumentParser(LocustTestCase):
     def test_parse_locustfile_empty_directory_error(self):
         with mock.patch("sys.stderr", new=StringIO()):
             with self.assertRaises(SystemExit):
-                locustfiles = parse_locustfile_option(
+                parse_locustfile_option(
                     args=[
                         "-f",
                         self.parent_dir.name,
@@ -298,7 +300,7 @@ class TestArgumentParser(LocustTestCase):
     def test_parse_locustfile_invalid_directory_error(self):
         with mock.patch("sys.stderr", new=StringIO()):
             with self.assertRaises(SystemExit):
-                locustfiles = parse_locustfile_option(
+                parse_locustfile_option(
                     args=[
                         "-f",
                         "non_existent_dir",
@@ -444,10 +446,10 @@ class TestFindLocustfiles(LocustTestCase):
                 with mock_locustfile() as mocked3:
                     locustfiles = parse_locustfile_paths([mocked1.file_path, mocked2.file_path])
 
-                    self.assertIn(mocked1.file_path, locustfiles)
-                    self.assertIn(mocked2.file_path, locustfiles)
+                self.assertIn(mocked1.file_path, locustfiles)
+                self.assertIn(mocked2.file_path, locustfiles)
 
-                    assert 2 == len(locustfiles)
+                assert 2 == len(locustfiles)
 
     def test_find_locustfiles_error_for_invalid_file_extension(self):
         with mock.patch("sys.stderr", new=StringIO()):
