@@ -35,12 +35,12 @@ function SwarmEditForm({ onSubmit, numUsers, spawnRate }: ISwarmForm) {
 
           <TextField
             defaultValue={spawnRate || 1}
-            label='Ramp Up (users started/second)'
+            label='Ramp up (users started/second)'
             name='spawnRate'
           />
 
           <Button size='large' type='submit' variant='contained'>
-            Update Swarm
+            Update
           </Button>
         </Box>
       </Form>
